+++ sdk/ai/azopenai/models.go
@@ -1479,16 +1479,6 @@ type ElasticsearchIndexFieldMappingOptions struct {
 	VectorFields []string
 }
 
-// EmbeddingItem - Representation of a single embeddings relatedness comparison.
-type EmbeddingItem struct {
-	// REQUIRED; List of embeddings value for the input prompt. These represent a measurement of the vector-based relatedness
-	// of the provided input.
-	Embedding []float32
-
-	// REQUIRED; Index of the prompt to which the EmbeddingItem corresponds.
-	Index *int32
-}
-
 // Embeddings - Representation of the response data from an embeddings request. Embeddings measure the relatedness of text
 // strings and are commonly used for search, clustering, recommendations, and other similar
 // scenarios.
@@ -1508,6 +1498,12 @@ type EmbeddingsOptions struct {
 	// Unless you are embedding code, we suggest replacing newlines (\n) in your input with a single space, as we have observed
 	// inferior results when newlines are present.
 	Input []string
+
+	// The number of dimensions the resulting output embeddings should have. Only supported in text-embedding-3 and later models.
+	Dimensions *int32
+
+	// The response encoding format to use for embedding data.
+	EncodingFormat *EmbeddingEncodingFormat
 
 	// When using Azure OpenAI, specifies the input type to use for embedding search.
 	InputType *string