+++ sdk/ai/azopenai/models_serde.go
@@ -2929,7 +2929,7 @@ func (e *EmbeddingItem) UnmarshalJSON(data []byte) error {
 		var err error
 		switch key {
 		case "embedding":
-			err = unpopulate(val, "Embedding", &e.Embedding)
+			err = deserializeEmbeddingsArray(val, e)
 			delete(rawMsg, key)
 		case "index":
 			err = unpopulate(val, "Index", &e.Index)
@@ -2976,6 +2976,8 @@ func (e *Embeddings) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type EmbeddingsOptions.
 func (e EmbeddingsOptions) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
+	populate(objectMap, "dimensions", e.Dimensions)
+	populate(objectMap, "encoding_format", e.EncodingFormat)
 	populate(objectMap, "input", e.Input)
 	populate(objectMap, "input_type", e.InputType)
 	populate(objectMap, "model", e.DeploymentName)
@@ -2992,6 +2994,12 @@ func (e *EmbeddingsOptions) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
+		case "dimensions":
+			err = unpopulate(val, "Dimensions", &e.Dimensions)
+			delete(rawMsg, key)
+		case "encoding_format":
+			err = unpopulate(val, "EncodingFormat", &e.EncodingFormat)
+			delete(rawMsg, key)
 		case "input":
 			err = unpopulate(val, "Input", &e.Input)
 			delete(rawMsg, key)